# internal/auth/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.15.0 DO NOT EDIT.
package ports

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// SignIn defines model for SignIn.
type SignIn struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// SignUp defines model for SignUp.
type SignUp struct {
	Name     string `json:"name"`
	Password string `json:"password"`
	Username string `json:"username"`
}

// SignInResult defines model for SignInResult.
type SignInResult struct {
	Status string  `json:"status"`
	Token  *string `json:"token,omitempty"`
}

// SignUpResult defines model for SignUpResult.
type SignUpResult struct {
	Status string `json:"status"`
	UserId *int64 `json:"userId,omitempty"`
}

// UnexpectedError defines model for UnexpectedError.
type UnexpectedError = Error

// SignInJSONRequestBody defines body for SignIn for application/json ContentType.
type SignInJSONRequestBody = SignIn

// SignUpJSONRequestBody defines body for SignUp for application/json ContentType.
type SignUpJSONRequestBody = SignUp