# internal/users/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.15.0 DO NOT EDIT.
package ports

import (
	"time"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// User defines model for User.
type User struct {
	Id               string     `json:"id"`
	Karma            int        `json:"karma"`
	LastLogin        *time.Time `json:"lastLogin,omitempty"`
	Name             string     `json:"name"`
	PostsCount       int        `json:"postsCount"`
	RegistrationDate time.Time  `json:"registrationDate"`
}

// UserArray defines model for UserArray.
type UserArray = []User

// FollowStatus defines model for FollowStatus.
type FollowStatus struct {
	Status string `json:"status"`
}

// UnexpectedError defines model for UnexpectedError.
type UnexpectedError = Error

// UserData defines model for UserData.
type UserData = User

// UserFollowers defines model for UserFollowers.
type UserFollowers = UserArray

// UserFollows defines model for UserFollows.
type UserFollows = UserArray